fix: git commit using github API
Changes the way that we are creating GitHub trees, because we are hitting
some problems in the GitHub side receiving 502 error with no more information
that "Server Error" and this kind of errors are not documented.

This changes the way that we create the new tree in GitHub. Now, instead
of creating the full tree to commit, we creates a new tree with only the
changes and referring to the tree of the previous commit.